+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- todo: add entry about change to default provider behavior
+
 # 12.14.0
 - [added] Added `AppAttestProviderFactory` to simplify App Check setup when
   using the App Attest provider. (#16182)